Cascade the long doc-id space across modules
Follow the widened ScoreDoc/reader APIs through the modules, and handle
features gracefully at >2^31 rather than throwing. Grouping splits its
group heads into one FixedBitSet per leaf (was a single global bit set) so
they stay int-addressable; the facet, suggest, spellcheck, classification
and function-query consumers that fed a composite maxDoc()/numDocs() into
scoring now use totalMaxDoc()/totalNumDocs(). Leaf-local conversions use a
plain cast (global doc minus its leaf docBase fits an int). Taxonomy
ordinals and the single-doc monitor matcher stay int (genuinely bounded).

+++ lucene/analysis/common/src/java/org/apache/lucene/analysis/synonym/word2vec/Word2VecSynonymProvider.java
@@ -92,7 +92,7 @@ public class Word2VecSynonymProvider {
 
       for (int i = 0; i < topDocs.scoreDocs.length; i++) {
         float similarity = topDocs.scoreDocs[i].score;
-        int id = topDocs.scoreDocs[i].doc;
+        int id = Math.toIntExact(topDocs.scoreDocs[i].doc);
 
         BytesRef synonym = word2VecModel.termValue(id);
         // We remove the original query term
