+++ lucene/facet/src/test/org/apache/lucene/facet/TestDrillSideways.java
@@ -1390,7 +1390,7 @@ public class TestDrillSideways extends FacetTestCase {
         DrillSidewaysResult dsr = ds.search(ddq, filter, null, numDocs, sort, true);
         actual.resultCount = dsr.hits.totalHits;
         for (ScoreDoc scoreDoc : dsr.hits.scoreDocs) {
-          actual.results.add(new DocAndScore(scoreDoc.doc, scoreDoc.score));
+          actual.results.add(new DocAndScore(Math.toIntExact(scoreDoc.doc), scoreDoc.score));
         }
         actual.facets = dsr.facets;
       } else {
@@ -1580,7 +1580,7 @@ public class TestDrillSideways extends FacetTestCase {
     @Override
     public LeafCollector getLeafCollector(LeafReaderContext context) throws IOException {
 
-      int docBase = context.docBase;
+      int docBase = Math.toIntExact(context.docBase);
       SortedDocValues idVals = DocValues.getSorted(context.reader(), "id");
 
       return new SimpleLeafCollector() {
