+++ lucene/grouping/src/test/org/apache/lucene/search/grouping/TestGrouping.java
@@ -902,7 +902,7 @@ public class TestGrouping extends LuceneTestCase {
 
       docStarts = new int[subSearchers.length];
       for (int subIDX = 0; subIDX < docStarts.length; subIDX++) {
-        docStarts[subIDX] = leaves.get(subIDX).docBase;
+        docStarts[subIDX] = Math.toIntExact(leaves.get(subIDX).docBase);
         // System.out.println("docStarts[" + subIDX + "]=" + docStarts[subIDX]);
       }
     }
@@ -1066,7 +1066,7 @@ public class TestGrouping extends LuceneTestCase {
         final ScoreDoc[] hits =
             s.search(new TermQuery(new Term("content", "real" + contentID)), numDocs).scoreDocs;
         for (ScoreDoc hit : hits) {
-          int idValue = docIDToID[hit.doc];
+          int idValue = docIDToID[Math.toIntExact(hit.doc)];
 
           final GroupDoc gd = groupDocs[idValue];
           seenIDs.add(idValue);
@@ -1123,10 +1123,10 @@ public class TestGrouping extends LuceneTestCase {
             sBlocks.search(new TermQuery(new Term("content", "real" + contentID)), numDocs)
                 .scoreDocs;
         for (ScoreDoc hit : hits) {
-          final GroupDoc gd = groupDocsByID[docIDToIDBlocks[hit.doc]];
+          final GroupDoc gd = groupDocsByID[docIDToIDBlocks[Math.toIntExact(hit.doc)]];
           assertTrue(gd.score2 == 0.0);
           gd.score2 = hit.score;
-          assertEquals(gd.id, docIDToIDBlocks[hit.doc]);
+          assertEquals(gd.id, docIDToIDBlocks[Math.toIntExact(hit.doc)]);
           // System.out.println("    score=" + gd.score + " score2=" + hit.score + " id=" +
           // docIDToIDBlocks[hit.doc]);
           termScoreMap.put(gd.score, gd.score2);
@@ -1380,7 +1380,8 @@ public class TestGrouping extends LuceneTestCase {
                       + " totalHits="
                       + gd.totalHits().value());
               for (ScoreDoc sd : gd.scoreDocs()) {
-                System.out.println("    id=" + docIDToID[sd.doc] + " score=" + sd.score);
+                System.out.println(
+                    "    id=" + docIDToID[Math.toIntExact(sd.doc)] + " score=" + sd.score);
               }
             }
 
@@ -1405,7 +1406,8 @@ public class TestGrouping extends LuceneTestCase {
                       + " totalHits="
                       + gd.totalHits().value());
               for (ScoreDoc sd : gd.scoreDocs()) {
-                System.out.println("    id=" + docIDToID[sd.doc] + " score=" + sd.score);
+                System.out.println(
+                    "    id=" + docIDToID[Math.toIntExact(sd.doc)] + " score=" + sd.score);
               }
             }
           }
@@ -1469,7 +1471,8 @@ public class TestGrouping extends LuceneTestCase {
                       + " totalHits="
                       + gd.totalHits().value());
               for (ScoreDoc sd : gd.scoreDocs()) {
-                System.out.println("    id=" + docIDToIDBlocks[sd.doc] + " score=" + sd.score);
+                System.out.println(
+                    "    id=" + docIDToIDBlocks[Math.toIntExact(sd.doc)] + " score=" + sd.score);
                 if (first) {
                   System.out.println("explain: " + sBlocks.explain(query, sd.doc));
                   first = false;
@@ -1498,7 +1501,7 @@ public class TestGrouping extends LuceneTestCase {
           // Fixup scores for reader2
           for (GroupDocs<?> groupDocsHits : expectedGroups.groups) {
             for (ScoreDoc hit : groupDocsHits.scoreDocs()) {
-              final GroupDoc gd = groupDocsByID[hit.doc];
+              final GroupDoc gd = groupDocsByID[Math.toIntExact(hit.doc)];
               assertEquals(gd.id, hit.doc);
               // System.out.println("fixup score " + hit.score + " to " + gd.score2 + " vs " +
               // gd.score);
@@ -1556,7 +1559,7 @@ public class TestGrouping extends LuceneTestCase {
         final ScoreDoc sd = group.scoreDocs()[hitIDX];
         assertEquals(
             "doc=" + sd.doc + " wrong shard",
-            ReaderUtil.subIndex(sd.doc, docStarts),
+            ReaderUtil.subIndex(Math.toIntExact(sd.doc), docStarts),
             sd.shardIndex);
       }
     }
@@ -1768,7 +1771,7 @@ public class TestGrouping extends LuceneTestCase {
         final FieldDoc actualFD = (FieldDoc) actualFDs[docIDX];
         // System.out.println("  actual doc=" + docIDtoID[actualFD.doc] + " score=" +
         // actualFD.score);
-        assertEquals(expectedFD.doc, docIDtoID[actualFD.doc]);
+        assertEquals(expectedFD.doc, docIDtoID[Math.toIntExact(actualFD.doc)]);
         assertArrayEquals(expectedFD.fields, actualFD.fields);
       }
     }
