+++ lucene/highlighter/src/java/org/apache/lucene/search/matchhighlight/MatchRegionRetriever.java
@@ -216,7 +216,10 @@ public class MatchRegionRetriever {
   public void highlightDocuments(TopDocs topDocs, MatchOffsetsConsumer consumer)
       throws IOException {
     highlightDocuments(
-        Arrays.stream(topDocs.scoreDocs).mapToInt(scoreDoc -> scoreDoc.doc).sorted().iterator(),
+        Arrays.stream(topDocs.scoreDocs)
+            .mapToInt(scoreDoc -> Math.toIntExact(scoreDoc.doc))
+            .sorted()
+            .iterator(),
         consumer,
         _ -> Integer.MAX_VALUE);
   }
@@ -362,7 +365,7 @@ public class MatchRegionRetriever {
         reader = currentContext.reader();
       }
 
-      int contextRelativeDocId = docId - currentContext.docBase;
+      int contextRelativeDocId = (int) (docId - currentContext.docBase);
 
       var fieldVisitor = new StoredFieldsVisitor(shouldLoadStoredField);
       StoredFields storedFields = reader.storedFields();
